+++ django_mysql/models/fields/sets.py
@@ -156,6 +156,7 @@ class SetCharField(SetFieldMixin, CharField):
         has_base_error = any(e.id == 'django_mysql.E001' for e in errors)
         if (
             not has_base_error and
+            self.max_length is not None and
             isinstance(self.base_field, CharField) and
             self.size
         ):